+++ Changelog.md
@@ -3,6 +3,7 @@
 - Add --delete parameter
 - Change progress message
 - Fix --version parameter
+- Add time and bandwidth to stats
 
 
 # v0.0.1